Move class definitions inside the folder to keep top-level .rb file only with requires
